CEXT-6527: relocate domain steps to management/domains
