+++ packages/aio-commerce-lib-app/source/management/domains/custom-installation/define.ts
