+++ packages/aio-commerce-lib-app/source/management/domains/events/utils.ts
