+++ packages/aio-commerce-lib-app/test/fixtures/eventing.ts
@@ -20,11 +20,11 @@ import type {
   CustomAdobeIoEventsApiClient,
   CustomCommerceEventsApiClient,
   EventsExecutionContext,
-} from "#management/installation/events/context";
+} from "#management/domains/events/context";
 import type {
   ExistingCommerceEventingData,
   ExistingIoEventsData,
-} from "#management/installation/events/utils";
+} from "#management/domains/events/utils";
 
 /** Creates a mock {@link EventProvider} with the given label and optional key. */
 export function createMockProvider(label: string, key?: string): EventProvider {
@@ -330,18 +330,18 @@ export async function importCommerceEventsStepWithMocks() {
     setSystemConfigByKey: vi.fn().mockResolvedValue(undefined),
   };
 
-  vi.doMock("#management/installation/events/helpers", async () => {
+  vi.doMock("#management/domains/events/helpers", async () => {
     const actual = await vi.importActual<
-      typeof import("#management/installation/events/helpers")
-    >("#management/installation/events/helpers");
+      typeof import("#management/domains/events/helpers")
+    >("#management/domains/events/helpers");
 
     return { ...actual, ...helperMocks };
   });
 
-  vi.doMock("#management/installation/events/utils", async () => {
+  vi.doMock("#management/domains/events/utils", async () => {
     const actual = await vi.importActual<
-      typeof import("#management/installation/events/utils")
-    >("#management/installation/events/utils");
+      typeof import("#management/domains/events/utils")
+    >("#management/domains/events/utils");
 
     return { ...actual, ...utilsMocks };
   });
@@ -354,9 +354,7 @@ export async function importCommerceEventsStepWithMocks() {
     return { ...actual, ...configMocks };
   });
 
-  const commerceModule = await import(
-    "#management/installation/events/commerce"
-  );
+  const commerceModule = await import("#management/domains/events/commerce");
 
   return {
     commerceEventsStep: commerceModule.commerceEventsStep,
@@ -370,7 +368,7 @@ export async function importCommerceEventsStepWithMocks() {
 export function cleanupCommerceEventsStepMocks() {
   vi.clearAllMocks();
   vi.resetModules();
-  vi.doUnmock("#management/installation/events/helpers");
-  vi.doUnmock("#management/installation/events/utils");
+  vi.doUnmock("#management/domains/events/helpers");
+  vi.doUnmock("#management/domains/events/utils");
   vi.doUnmock("@adobe/aio-commerce-lib-config");
 }
