+++ packages/aio-commerce-lib-app/test/unit/management/domains/events/commerce.test.ts
@@ -13,7 +13,7 @@
 import { afterEach, describe, expect, test, vi } from "vitest";
 
 import { isLeafStep } from "#management/common/workflow/step";
-import { commerceEventsStep } from "#management/installation/events/commerce";
+import { commerceEventsStep } from "#management/domains/events/commerce";
 import {
   configWithCommerceEventing,
   configWithExternalEventing,
